avoid fetching all applications on single application routes
