+++ src/routes/(pages)/app/(app)/rpgf/rounds/[slugOrId]/applications/+layout.ts
@@ -1,13 +1,10 @@
-import { getApplications } from '$lib/utils/rpgf/rpgf.js';
 import type { InProgressBallot } from '$lib/utils/rpgf/types/ballot';
 import storedWritable from '@efstajas/svelte-stored-writable';
 import { error } from '@sveltejs/kit';
 import { z } from 'zod';
 
-export const load = async ({ parent, route, url, depends, fetch }) => {
-  depends('rpgf:round:applications');
-
-  const { round, existingBallot, rpgfUserData } = await parent();
+export const load = async ({ parent, route }) => {
+  const { round, existingBallot } = await parent();
 
   if (!round.published) {
     throw error(404);
@@ -30,39 +27,6 @@ export const load = async ({ parent, route, url, depends, fetch }) => {
     existingBallot?.ballot ?? {},
   );
 
-  const { resultsPublished } = round;
-
-  const sortByParam: string =
-    url.searchParams.get('sortBy') ?? (resultsPublished ? 'allocation' : 'createdAt');
-  const filterParam: string | null = url.searchParams.get('filter');
-
-  const allApplications =
-    filterParam === 'own' && !rpgfUserData
-      ? []
-      : await getApplications(
-          fetch,
-          round.id,
-          100000, // Fetch all applications, later we can paginate
-          0,
-          (() => {
-            switch (sortByParam) {
-              case 'createdAt':
-                return 'createdAt:desc';
-              case 'name':
-                return 'name:asc';
-              case 'allocation':
-                return 'allocation:desc';
-              default:
-                return undefined;
-            }
-          })(),
-          filterParam === 'own' && rpgfUserData ? rpgfUserData.userId : null,
-          filterParam === 'approved' || filterParam === 'rejected' || filterParam === 'pending'
-            ? filterParam
-            : undefined,
-          filterParam?.startsWith('cat-') ? filterParam.replaceAll('cat-', '') : undefined,
-        );
-
   // If true, display sidebar that lets admins review applications
   const reviewMode =
     route.id !== '/(pages)/app/(app)/rpgf/rounds/[slugOrId]/applications/new' &&
@@ -73,9 +37,6 @@ export const load = async ({ parent, route, url, depends, fetch }) => {
     reviewMode,
     voteMode,
     ballot: ballotWritable,
-    allApplications,
-    sortByParam,
-    filterParam,
     resultsMode,
   };
 };
